+++ stepmania/src/RageTypes.cpp
@@ -26,13 +26,13 @@ void RageColor::FromStack( lua_State *L, int iPos )
 	int iFrom = lua_gettop( L );
 
 	lua_rawgeti( L, iFrom, 1 );
-	r = lua_tonumber( L, -1 );
+	r = (float)lua_tonumber( L, -1 );
 	lua_rawgeti( L, iFrom, 2 );
-	g = lua_tonumber( L, -1 );
+	g = (float)lua_tonumber( L, -1 );
 	lua_rawgeti( L, iFrom, 3 );
-	b = lua_tonumber( L, -1 );
+	b = (float)lua_tonumber( L, -1 );
 	lua_rawgeti( L, iFrom, 4 );
-	a = lua_tonumber( L, -1 );
+	a = (float)lua_tonumber( L, -1 );
 	lua_pop( L, 5 );
 }
